perf(e2e): extract API tests and reduce wait times
- Extract 15 API-only tests into api.spec.ts (run once on chromium)
- Keep 3-browser sequential chain for browser-specific tests
- Browser tests reduced from 45 to 30 per engine

Addresses #233
