+++ src/frontend/e2e-tests/e2e/helpers.ts
@@ -38,15 +38,7 @@ export async function loginViaUI(page: Page, email: string, password: string) {
   const cx = width / 2;
   const f = fv(page);
 
-  // The "Enable accessibility" button (if present) can cover the center of
-  // the screen and intercept our login field clicks. Dismiss it first if visible.
-  const accessibilityBtn = page.locator("button", {
-    hasText: "Enable accessibility",
-  });
-  if (await accessibilityBtn.isVisible({ timeout: 500 }).catch(() => false)) {
-    await accessibilityBtn.click();
-    await page.waitForTimeout(300);
-  }
+  await dismissAccessibility(page);
 
   await f.click({ position: { x: cx, y: height * 0.46 }, force: true });
   await page.waitForTimeout(200);
@@ -70,13 +62,7 @@ export async function tryLogin(page: Page, email: string, password: string) {
   const cx = width / 2;
   const f = fv(page);
 
-  const accessibilityBtn = page.locator("button", {
-    hasText: "Enable accessibility",
-  });
-  if (await accessibilityBtn.isVisible({ timeout: 500 }).catch(() => false)) {
-    await accessibilityBtn.click();
-    await page.waitForTimeout(300);
-  }
+  await dismissAccessibility(page);
 
   await f.click({ position: { x: cx, y: height * 0.46 }, force: true });
   await page.waitForTimeout(200);
@@ -102,6 +88,16 @@ export async function waitForFlutter(page: Page) {
   // Wait for flutter-view to be present and rendered
   await page.waitForSelector("flutter-view", { timeout: 10_000 });
   await page.waitForTimeout(500);
+}
+
+/** Dismiss the "Enable accessibility" button if visible. Flutter shows this
+ *  overlay on each route load and it can cover the canvas / terminal. */
+async function dismissAccessibility(page: Page) {
+  const btn = page.locator("button", { hasText: "Enable accessibility" });
+  if (await btn.isVisible({ timeout: 500 }).catch(() => false)) {
+    await btn.click();
+    await page.waitForTimeout(300);
+  }
 }
 
 export function fv(page: Page) {
@@ -349,6 +345,7 @@ export async function openWorkspace(
   await waitForFlutter(page);
   await containerPromise;
   await terminalPromise;
+  await dismissAccessibility(page);
 }
 
 /** Convenience: register user, create workspace, open it. */
