+++ src/frontend/e2e-tests/playwright.config.ts
@@ -37,8 +37,10 @@ const webkitUse = {
   browserName: "webkit" as const,
 };
 
-// Browsers run sequentially (chromium → firefox → webkit) to avoid
-// overwhelming SQLite with concurrent writes from parallel tests.
+// Test projects:
+// - chromium-api: API-only + simple UI tests that don't need cross-browser (run once)
+// - chromium → firefox → webkit: browser-specific tests (sequential to avoid
+//   container startup contention from parallel browser projects)
 
 export default defineConfig({
   testDir: "./e2e",
@@ -58,6 +60,12 @@ export default defineConfig({
     screenshot: "only-on-failure",
   },
   projects: [
+    {
+      // API-only and simple-UI tests — no cross-browser behavior, run once.
+      name: "chromium-api",
+      testMatch: "api.spec.ts",
+      use: chromiumUse,
+    },
     {
       name: "chromium",
       testMatch: ["klangk.spec.ts", "terminal-keymap.spec.ts"],
